fix(gradle): Detect merge commits based on the parents instead of message
